% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsne_parse.R
\name{rtsne_convert_dataframe}
\alias{rtsne_convert_dataframe}
\title{Title}
\usage{
rtsne_convert_dataframe(
  rtnse_fit,
  ids,
  sample_metadata_dataframe = NULL,
  id_column_in_metadata = "id"
)
}
\arguments{
\item{rtnse_fit}{object returned by \link[Rtsne]{Rtsne}}

\item{ids}{name of each samples input into Rtsne (character).}

\item{sample_metadata_dataframe}{sample metadata. Should be 1 row per sample (dataframe)}

\item{id_column_in_metadata}{name of column in sample_metadata_dataframe containing sample identifier (string)}
}
\value{
dataframe
}
\description{
Title
}
