% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RtsneWide.R
\name{tsne_long}
\alias{tsne_long}
\title{Run tsne from long data format}
\usage{
tsne_long(data, id_col, names_from, values_from, check_duplicates = TRUE, ...)
}
\arguments{
\item{data}{long format dataframe - should contain one column with sample level identifiers (data.frame)}

\item{id_col}{name of column describing ID (string)}

\item{names_from}{column containing names of the features (string)}

\item{values_from}{column containing values of the feature}

\item{check_duplicates}{Should Rtsne check for dups after PCA (bool)}

\item{...}{
  Arguments passed on to \code{\link[=tsne_wide]{tsne_wide}}
  \describe{
    \item{\code{perplexity}}{tsne perplexity}
    \item{\code{remove_non_numeric_columns}}{remove non numeric columns (bool)}
    \item{\code{remove_zero_variance_columns}}{remove columns with zero variance (bool)}
    \item{\code{return_preprocessed_dt}}{Don't run TSNE. Just preprocess the data then return the data.table bool}
    \item{\code{input_is_mutable_datatable}}{are we allowed to mutate the input dataset in place (Should never be directly set - defaults to false) (bool)}
  }}
}
\value{
Rtsne dataframe
}
\description{
Run tsne from long data format
}
\details{
... param also allows user to set any param of \link[Rtsne]{Rtsne}
}
