% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RtsneWide.R
\name{tsne_wide}
\alias{tsne_wide}
\title{Title}
\usage{
tsne_wide(
  data,
  id_col = NULL,
  perplexity = 30,
  remove_non_numeric_columns = TRUE,
  remove_zero_variance_columns = TRUE,
  check_duplicates = TRUE,
  input_is_mutable_datatable = FALSE,
  return_preprocessed_dt = FALSE,
  ...
)
}
\arguments{
\item{data}{data to run TSNE on. Each row should be a sample. Must contain at leat one numeric column}

\item{id_col}{name of the column in data that represents the sample identifier.}

\item{perplexity}{tsne perplexity}

\item{remove_non_numeric_columns}{remove non numeric columns (bool)}

\item{remove_zero_variance_columns}{remove columns with zero variance (bool)}

\item{check_duplicates}{Should Rtsne check for dups after PCA (bool)}

\item{input_is_mutable_datatable}{are we allowed to mutate the input dataset in place (Should never be directly set - defaults to false) (bool)}

\item{return_preprocessed_dt}{Don't run TSNE. Just preprocess the data then return the data.table bool}

\item{...}{set any other argument of \link[Rtsne]{Rtsne}}
}
\value{
Rtsne (tibble)
}
\description{
Title
}
\examples{
\dontrun{
tsne_wide(iris)
}
}
